+++ CHANGELOG.md
@@ -28,6 +28,7 @@ This section is for maintaining a changelog for all breaking changes for the cli
 ### Fixed
 - Fix version and build ([#254](https://github.com/opensearch-project/opensearch-java/pull/254))
 - Fix integer overflow for variables in indices stats response ([#960](https://github.com/opensearch-project/opensearch-java/pull/960))
+- Fix composite aggregations for search requests ([#967](https://github.com/opensearch-project/opensearch-java/pull/967))
 
 
 ### Security